test(racingcar) : 리팩토링 사항에 맞춰 자동차 테스트 코드 변경
