+++ src/test/java/racingcar/CarTest.java
@@ -1,38 +1,39 @@
 package racingcar;
 
-import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
+import static camp.nextstep.edu.missionutils.test.Assertions.assertRandomNumberInRangeTest;
+import static org.assertj.core.api.Assertions.assertThat;
 
-import org.junit.jupiter.api.BeforeEach;
+import camp.nextstep.edu.missionutils.test.NsTest;
 import org.junit.jupiter.api.Test;
 
-class CarTest {
-    private Car pobi;
-    private Car woni;
-    private Car jun;
-    @BeforeEach
-    void setUp() {
-        pobi = new Car("pobi");
-        woni = new Car("woni");
-        jun = new Car("jun");
+class CarTest extends NsTest {
+    private static final int MOVING_FORWARD = 4;
+    private static final int STOP = 3;
+
+    public static void main(String[] args) {
+        Car pobi = new Car("pobi");
+        Car woni = new Car("woni");
+
+        pobi.moveIfSatisfy(2);
+        woni.moveIfSatisfy(2);
+
+        System.out.println(pobi);
+        System.out.println(woni);
     }
 
     @Test
-    void Car_출력_테스트() {
-        //given
-        //when
-        pobi.move();
-        pobi.move();
-        woni.move();
-        woni.move();
-        woni.move();
-        woni.move();
-        jun.move();
-        jun.move();
-        jun.move();
-        //then
-        assertThat(pobi.toString()).isEqualTo("pobi : --");
-        assertThat(woni.toString()).isEqualTo("woni : ----");
-        assertThat(jun.toString()).isEqualTo("jun : ---");
+    void 전진_정지() {
+        assertRandomNumberInRangeTest(
+                () -> {
+                    main(new String[]{});
+                    assertThat(output()).contains("pobi : -", "woni : --"); // output() 메서드의 trim() 적용됨을 유의
+                },
+                MOVING_FORWARD, STOP, MOVING_FORWARD, MOVING_FORWARD
+        );
+    }
+
+    @Override
+    public void runMain() {
+        main(new String[]{});
     }
 }
